fpcrawler/requirements.txt
aiohttp==3.5.4
aioredis==1.2.0
amqp==2.4.1
arelle==2.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
crochet==1.9.0
cryptography==2.5
cssselect==1.0.3
daphne==2.2.5
dart-fss==0.1.12
dataclasses==0.6
Django==2.1.7
djangorestframework==3.9.1
djongo==1.2.31
ephem==3.7.6.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
finance-datareader==0.7.2
hiredis==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
idna-ssl==1.1.0
incremental==17.5.0
isodate==0.6.0
jdcal==1.4.1
kombu==4.3.0
log4mongo==1.6.2
lunardate==0.2.0
lxml==4.3.0
more-itertools==5.0.0
msgpack==0.6.1
multidict==4.5.2
nose==1.3.7
numpy==1.16.1
openpyxl==2.6.2
pandas==0.24.1
parsel==1.5.1
pluggy==0.8.1
psycopg2==2.7.7
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pyCalverter==1.6.1
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyluach==0.3.0.dev1
pymongo==3.7.2
pyOpenSSL==19.0.0
pytest==4.2.0
pytest-asyncio==0.10.0
pytest-django==3.4.7
python-dateutil==2.8.0
python-logstash==0.4.6
pytz==2018.9
queuelib==1.5.0
redis==3.1.0
requests==2.21.0
requests-file==1.4.3
Scrapy==1.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.7.3
sqlparse==0.2.4
tqdm==4.33.0
Twisted==18.9.0
txaio==18.8.1
typing-extensions==3.7.2
urllib3==1.24.1
vine==1.2.0
w3lib==1.20.0
webencodings==0.5.1
workalendar==4.2.0
wrapt==1.11.1
xlrd==1.2.0
yarl==1.3.0
zope.interface==4.6.0
